Complete pep8 renaming for weather subpackage
