--- reskit/weather/Era5Source/__init__.py
@@ -1,1 +1,0 @@
-from reskit.weather.Era5Source import Era5Source
