--- reskit/weather/GSA_mean/__init__.py
@@ -1,1 +1,0 @@
-from .GSAmeanSource import GSAmeanSource
