+++ reskit/weather/__init__.py
@@ -3,10 +3,10 @@ from reskit.weather import util
 from reskit.weather.nc_source import NCSource
 
 from reskit.weather.merra_source.merra_source import MerraSource
-from reskit.weather.Era5Source.Era5Source import Era5Source
+from reskit.weather.era_5_source.era_5_source import Era5Source
 from reskit.weather.sarah_source import SarahSource
-from reskit.weather.GSA_mean.GSAmeanSource import GSAmeanSource
-from reskit.weather.GWA_mean.GWAmeanSource import GWAmeanSource
+from reskit.weather.gsa_mean.gsa_mean_source import GSAmeanSource
+from reskit.weather.gwa_mean.gwa_mean_source import GWAmeanSource
 from reskit.weather.IconlamSource.IconlamSource import IconlamSource
 
 # from .CosmoSource import CosmoSource  # TODO: NEEDS UPDATING!
