+++ reskit/weather/cosmo_source/__init__.py
@@ -1,1 +1,1 @@
-from .CosmoSource import CosmoSource
+from reskit.weather.cosmo_source.cosmo_source import CosmoSource
