+++ reskit/weather/era_5_source/data/ERA5_DNI_mean_archive.tif
