+++ reskit/weather/era_5_source/data/ERA5_DNI_percentile_90_2000_to_2020.tif
