+++ reskit/weather/era_5_source/data/ERA5_surface_solar_radiation_downwards_mean.tiff
