+++ reskit/weather/era_5_source/data/ERA5_total_sky_direct_solar_radiation_at_surface_mean.tiff
