+++ reskit/weather/era_5_source/data/ERA5_wind_direction_100m_mean.tiff
