+++ reskit/weather/gsa_mean/__init__.py
@@ -1,0 +1,1 @@
+from reskit.weather.gsa_mean.gsa_mean_source import GSAmeanSource
